merge new features

+++ modules/perception/common/algorithm/geometry/convex_hull_2d.h
@@ -159,21 +159,22 @@ bool ConvexHull2D<CLOUD_IN_TYPE, CLOUD_OUT_TYPE>::MockConvexHull(
     }
   }
 
-  out_polygon->at(0).x = static_cast<float>(minv(0));
-  out_polygon->at(0).y = static_cast<float>(minv(1));
-  out_polygon->at(0).z = static_cast<float>(minv(2));
+  // double NOT NECESSARY to static_cast float
+  out_polygon->at(0).x = minv(0);
+  out_polygon->at(0).y = minv(1);
+  out_polygon->at(0).z = minv(2);
 
-  out_polygon->at(1).x = static_cast<float>(maxv(0));
-  out_polygon->at(1).y = static_cast<float>(minv(1));
-  out_polygon->at(1).z = static_cast<float>(minv(2));
+  out_polygon->at(1).x = maxv(0);
+  out_polygon->at(1).y = minv(1);
+  out_polygon->at(1).z = minv(2);
 
-  out_polygon->at(2).x = static_cast<float>(maxv(0));
-  out_polygon->at(2).y = static_cast<float>(maxv(1));
-  out_polygon->at(2).z = static_cast<float>(minv(2));
+  out_polygon->at(2).x = maxv(0);
+  out_polygon->at(2).y = maxv(1);
+  out_polygon->at(2).z = minv(2);
 
-  out_polygon->at(3).x = static_cast<float>(minv(0));
-  out_polygon->at(3).y = static_cast<float>(maxv(1));
-  out_polygon->at(3).z = static_cast<float>(minv(2));
+  out_polygon->at(3).x = minv(0);
+  out_polygon->at(3).y = maxv(1);
+  out_polygon->at(3).z = minv(2);
   return true;
 }
 
@@ -226,13 +227,14 @@ bool ConvexHull2D<CLOUD_IN_TYPE, CLOUD_OUT_TYPE>::GetConvexHullMonotoneChain(
   }
   out_polygon->clear();
   out_polygon->resize(polygon_indices_.size());
-  float min_z = static_cast<float>(in_cloud_->at(0).z);
+  // double NOT NECESSARY to static_cast float
+  double min_z = in_cloud_->at(0).z;
   for (std::size_t id = 0; id < in_cloud_->size(); ++id) {
-    min_z = std::min<float>(static_cast<float>(in_cloud_->at(id).z), min_z);
+    min_z = std::min<double>(in_cloud_->at(id).z, min_z);
   }
   for (std::size_t i = 0; i < polygon_indices_.size(); ++i) {
-    out_polygon->at(i).x = static_cast<float>(points_[polygon_indices_[i]](0));
-    out_polygon->at(i).y = static_cast<float>(points_[polygon_indices_[i]](1));
+    out_polygon->at(i).x = points_[polygon_indices_[i]](0);
+    out_polygon->at(i).y = points_[polygon_indices_[i]](1);
     out_polygon->at(i).z = min_z;
   }
   return true;
